Add TODOs and cleanup
Remove unused piyx2 code.

Change how getMainFeatures is used. Don't do work in
constructor. Create a get method and have it throw exceptions if
something goes wrong.

Add TODOs in drivetrain subsystem for things that need to be
added. Much of it can be done before there is a chassis available. We
just won't know what exact values to put in. So make them up for
now...